Update `README.md` with newer information & fix
There are still some important things needing to be addressed, such as
the use of `locate` which isn't at all guaranteed to be there, not to
mention just relies upon the user having a recent database.

Still, this cleans it up a bit. I corrected a typo or two. Reworded for
clarity, and I added the newer help output for usage information.

I've also added some dependencies which you'd missed, and added a list
of packages needed in Debian- or Ubuntu-based distributions of Linux.